+++ src/test/java/subway/controller/StationControllerTest.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
